requirements.txt
fastapi==0.110.2
uvicorn==0.27.1
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
jinja2==3.1.3
aiofiles==23.2.1
sqlalchemy==1.4.50
pydantic==2.10.0
pydantic-core==2.27.0
typing-extensions==4.12.2
cerebras_cloud_sdk
python-dotenv==1.0.0
gunicorn==21.2.0
pymongo[srv]==4.6.1
dnspython==2.6.1
motor==3.3.2
aiohttp==3.9.3
pillow>=9.5.0
